Extract file view window state
